+++ include/al5d_cpp/JointAngle.hpp
@@ -2,9 +2,12 @@
 #ifndef AL5D_CPP_JOINTANGLE_HPP
 #define AL5D_CPP_JOINTANGLE_HPP
 
+// SYSTEM INCLUDES
+#include <cstdint>
+
 namespace al5d
 {
-    typedef int JointAngle; // TODO : change int to PulseWidth
+    typedef uint16_t JointAngle;
 }
 
 #endif // AL5D_CPP_JOINTANGLE_HPP
